Inf sig fix

+++ src/BinnedDiffusion_transform.cxx
@@ -73,8 +73,8 @@ struct generate_random {
         typename GeneratorPool::generator_type rand_gen2 = rand_pool2.get_state();
 
         for (int k = 0; k < samples; k++) {
-            double u1 = rand_gen1.urand64(range_min, range_max1) / range_max1; 
-            double u2 = rand_gen2.urand64(range_min, range_max2) / range_max2; 
+            double u1 = (double) rand_gen1.urand64(range_min, range_max1) / range_max1; 
+            double u2 = (double) rand_gen2.urand64(range_min, range_max2) / range_max2; 
             normals(i * samples + k)     = sqrt(-2*log(u1)) * cos(2*PI*u2);
             normals(i * samples + k + 1) = sqrt(-2*log(u1)) * sin(2*PI*u2);
         }
